requirements.txt
# ============================================================================
# SAD - Sistema de Apoio à Decisão para Monitoramento de Idosos com Obesidade
# Dependências Python
# ============================================================================

# ===== CORE: Análise de Dados =====
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ===== MACHINE LEARNING =====
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0  # Para lidar com dados desbalanceados
joblib>=1.3.0  # Serialização de modelos

# ===== GEOESPACIAL =====
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.4.0
folium>=0.14.0  # Mapas interativos
geopy>=2.3.0  # Geocodificação de endereços

# ===== VISUALIZAÇÃO =====
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ===== BANCO DE DADOS =====
psycopg2-binary>=2.9.0  # PostgreSQL adapter
SQLAlchemy>=2.0.0  # ORM
GeoAlchemy2>=0.14.0  # Extensão geoespacial para SQLAlchemy

# ===== WEB FRAMEWORK =====
streamlit>=1.28.0  # Dashboard interativo
streamlit-folium>=0.15.0  # Integração Streamlit + Folium

# Alternativa: Flask (caso prefira API REST)
# Flask>=3.0.0
# Flask-SQLAlchemy>=3.1.0
# Flask-CORS>=4.0.0

# ===== UTILIDADES =====
python-dotenv>=1.0.0  # Gerenciamento de variáveis de ambiente
pyyaml>=6.0  # Leitura de arquivos de configuração
openpyxl>=3.1.0  # Leitura/escrita de arquivos Excel
xlrd>=2.0.0  # Leitura de arquivos .xls antigos

# ===== JUPYTER (opcional, para análise exploratória) =====
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# ===== QUALIDADE DE CÓDIGO =====
# black>=23.0.0  # Formatação automática
# flake8>=6.0.0  # Linting
# pytest>=7.4.0  # Testes automatizados
# pytest-cov>=4.1.0  # Cobertura de testes

# ===== OUTROS =====
requests>=2.31.0  # Requisições HTTP (caso integre APIs externas)
tqdm>=4.66.0  # Barras de progresso
